pyproject.toml
[project]
name = "udemy-langchain"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=0.3.22",
    "langchain-community>=0.3.20",
    "langchain-huggingface>=0.1.2",
    "langchain-ollama>=0.3.0",
    "python-dotenv>=1.1.0",
]
